WIP - Default to using suggested cases
Abstract out seperate components from TransferOrderAccordion.tsx
Selecting the 'Case not listed.' row in the table enables the form to
select the case. What is not working is selecting a different row as
it does not enable the 'Approve' button.

[skip ci]

Jira ticket: CAMS-338
