+++ user-interface/src/data-verification/transfer/CaseTable.tsx
@@ -1,21 +1,27 @@
 import { CaseNumber } from '@/lib/components/CaseNumber';
 import { formatDate } from '@/lib/utils/datetime';
 import { CaseSummary } from '@common/cams/cases';
-import { SyntheticEvent, forwardRef, useImperativeHandle, useState } from 'react';
+import {
+  PropsWithChildren,
+  SyntheticEvent,
+  forwardRef,
+  useImperativeHandle,
+  useState,
+} from 'react';
 
 export type CaseTableImperative = {
   clearAllCheckboxes: () => void;
 };
 
-interface CaseTableProps {
+type CaseTableProps = PropsWithChildren & {
   id: string;
-  cases: Array<CaseSummary>;
-  onSelect?: (bCase: CaseSummary) => void;
+  cases: Array<CaseSummary | null>;
+  onSelect?: (bCase: CaseSummary | null) => void;
   displayDocket?: boolean;
-}
+};
 
 function _CaseTable(props: CaseTableProps, CaseTableRef: React.Ref<CaseTableImperative>) {
-  const { id, cases, onSelect } = props;
+  const { id, cases, onSelect, children } = props;
 
   const [selectedIdx, setSelectedIdx] = useState<number | null>(null);
 
@@ -40,15 +46,33 @@ function _CaseTable(props: CaseTableProps, CaseTableRef: React.Ref<CaseTableImpe
         <tr>
           {onSelect && <th scope="col">Select</th>}
           <th scope="col">Case Number</th>
+          <th scope="col">Court (Division)</th>
           <th scope="col">Case Title</th>
-          <th scope="col">SSN/EIN</th>
-          <th scope="col">Court</th>
-          <th scope="col">Case Filed</th>
           <th scope="col">Chapter</th>
+          <th scope="col">Filed Date</th>
+          <th scope="col">Tax ID</th>
         </tr>
       </thead>
       <tbody>
         {cases?.map((bCase, idx) => {
+          if (!bCase) {
+            return (
+              <tr key={'empty'} data-testid={'empty-row'}>
+                <th scope="col">
+                  <input
+                    type="radio"
+                    onChange={handleCaseSelection}
+                    value={idx}
+                    name="case-selection"
+                    data-testid={`suggested-cases-radio-empty`}
+                    checked={idx === selectedIdx}
+                    title={`case not listed`}
+                  ></input>
+                </th>
+                <td colSpan={6}>Case not listed.</td>
+              </tr>
+            );
+          }
           const taxId = bCase.debtor?.ssn || bCase.debtor?.taxId || '';
           const key = `${id}-row-${idx}`;
           return (
@@ -69,20 +93,21 @@ function _CaseTable(props: CaseTableProps, CaseTableRef: React.Ref<CaseTableImpe
               <td scope="row">
                 <CaseNumber caseId={bCase.caseId} />
               </td>
-              <td scope="row">{bCase.caseTitle}</td>
-              <td scope="row" className="text-no-wrap">
-                {taxId}
-              </td>
               <td scope="row">
                 {bCase.courtName} ({bCase.courtDivisionName})
               </td>
+              <td scope="row">{bCase.caseTitle}</td>
+              <td scope="row">{bCase.chapter}</td>
               <td scope="row" className="text-no-wrap">
                 {formatDate(bCase.dateFiled)}
               </td>
-              <td scope="row">{bCase.chapter}</td>
+              <td scope="row" className="text-no-wrap">
+                {taxId}
+              </td>
             </tr>
           );
         })}
+        {children}
       </tbody>
     </table>
   );
