+++ user-interface/src/data-verification/transfer/TransferConfirmationModal.tsx
@@ -1,0 +1,156 @@
+import { forwardRef, useImperativeHandle, useRef, useState } from 'react';
+import Modal from '@/lib/components/uswds/modal/Modal';
+import { OrderStatus } from '@common/cams/orders';
+import { ModalRefType } from '@/lib/components/uswds/modal/modal-refs';
+import { getCaseNumber } from '@/lib/utils/formatCaseNumber';
+
+interface ConfirmationModalProps {
+  id: string;
+  fromCaseId: string;
+  toCaseId?: string;
+  fromDivisionName: string;
+  toDivisionName?: string;
+  fromCourtName: string;
+  toCourtName?: string;
+  onCancel: () => void;
+  onConfirm: (status: OrderStatus, reason?: string) => void;
+}
+
+type ShowOptionParams = {
+  status: OrderStatus;
+};
+
+type ShowOptions = {
+  status: OrderStatus;
+  title: string;
+};
+
+type ConfirmationModalImperative = ModalRefType & {
+  show: (options: ShowOptionParams) => void;
+};
+
+function ConfirmationModalComponent(
+  props: ConfirmationModalProps,
+  ConfirmationModalRef: React.Ref<ConfirmationModalImperative>,
+) {
+  const {
+    id,
+    fromCaseId,
+    toCaseId,
+    fromDivisionName,
+    toDivisionName,
+    fromCourtName,
+    toCourtName,
+    onConfirm,
+  }: ConfirmationModalProps = props;
+
+  const modalRef = useRef<ModalRefType>(null);
+  const reasonRef = useRef<HTMLTextAreaElement>(null);
+  const [reason] = useState<string>('');
+  const [options, setOptions] = useState<ShowOptions>({
+    status: 'pending',
+    title: '',
+  });
+
+  function clearReason() {
+    if (reasonRef.current) reasonRef.current.value = '';
+  }
+
+  const actionButtonGroup = {
+    modalId: `confirmation-modal-${id}`,
+    modalRef: modalRef,
+    submitButton: {
+      label: options.title,
+      onClick: () => {
+        onConfirm(options.status, reasonRef.current?.value);
+      },
+      className: options.status === 'rejected' ? 'usa-button--secondary' : '',
+    },
+    cancelButton: {
+      label: 'Go back',
+      onClick: () => {
+        clearReason();
+        hide();
+      },
+    },
+  };
+
+  function show(options: ShowOptionParams) {
+    const title = options.status === 'approved' ? 'Approve' : 'Reject';
+
+    setOptions({
+      status: options.status,
+      title,
+    });
+
+    if (modalRef.current?.show) {
+      modalRef.current?.show({});
+    }
+  }
+
+  function hide() {
+    if (modalRef.current?.hide) {
+      modalRef.current?.hide({});
+    }
+  }
+
+  useImperativeHandle(ConfirmationModalRef, () => ({
+    show,
+    hide,
+  }));
+
+  return (
+    <Modal
+      ref={modalRef}
+      modalId={`confirm-modal-${id}`}
+      className="confirm-modal"
+      heading={`${options.title} case transfer?`}
+      data-testid={`confirm-modal-${id}`}
+      onClose={clearReason}
+      content={
+        <>
+          This will {options.status === 'approved' ? 'approve' : 'stop'} the transfer of case
+          <span className="transfer-highlight__span">{getCaseNumber(fromCaseId)}</span>
+          in
+          <span className="transfer-highlight__span">
+            {fromCourtName} ({fromDivisionName})
+          </span>
+          {toCaseId && (
+            <>
+              to case
+              <span className="transfer-highlight__span">{getCaseNumber(toCaseId)}</span>
+            </>
+          )}
+          {toCourtName && (
+            <>
+              in
+              <span className="transfer-highlight__span">
+                {toCourtName} ({toDivisionName})
+              </span>
+            </>
+          )}
+          .
+          {options.status === 'rejected' && (
+            <div>
+              <label htmlFor={`rejection-reason-${id}`} className="usa-label">
+                Reason for rejection
+              </label>
+              <div>
+                <textarea
+                  id={`rejection-reason-${id}`}
+                  data-testid={`rejection-reason-input-${id}`}
+                  ref={reasonRef}
+                  className="rejection-reason-input usa-textarea"
+                  defaultValue={reason}
+                ></textarea>
+              </div>
+            </div>
+          )}
+        </>
+      }
+      actionButtonGroup={actionButtonGroup}
+    ></Modal>
+  );
+}
+
+export const ConfirmationModal = forwardRef(ConfirmationModalComponent);
